for 'XML' and 'InMemory' modes, Dockerfile is run for 'DB' mode, docker-compose.yml is run

+++ startup.sh
@@ -1,24 +1,30 @@
 #!/bin/sh
 
-docker-compose -f docker/docker-compose.yml build
-
 mode="$1"
 clear="$2"
 
 if [ "$mode" = "DB" ]; then
-docker-compose -f docker/docker-compose.yml run --rm python python main.py --db handbook --user handbook_user --password 111111 --host postgres_db --port 5432
+  printf "\nDB mode started: data is saved in the database."
+  docker-compose -f docker/docker-compose.yml build
+  docker-compose -f docker/docker-compose.yml run --rm python python main.py --db handbook --user handbook_user --password 111111 --host postgres_db --port 5432
 fi
 
 if [ "$mode" = "XML" ]; then
-docker-compose -f docker/docker-compose.yml run --rm python python main.py --path /handbook.xml
+  printf "\nXML mode started: data is saved in an XML file."
+  docker build -f docker/Dockerfile -t handbook_image .
+  docker run -it --rm --name handbook_container handbook_image python main.py --path /handbook.xml
 fi
 
 if [ "$mode" = "" ] || [ "$mode" = "InMemory" ]; then
-docker-compose -f docker/docker-compose.yml run --rm python python main.py
+  printf "\nInMemory mode started: data is saved in internal memory."
+  docker build -f docker/Dockerfile -t handbook_image .
+  docker run -it --rm --name handbook_container handbook_image python main.py
 fi
 
-if [ "$clear" = "clear" ]; then
-docker-compose -f docker/docker-compose.yml down -v
-else
-docker-compose -f docker/docker-compose.yml stop
+if [ "$mode" = "DB" ]; then
+  if [ "$clear" = "clear" ]; then
+    docker-compose -f docker/docker-compose.yml down -v
+  else
+    docker-compose -f docker/docker-compose.yml stop
+  fi
 fi
